refactor(conversations): lift the chat-mode bubbles into ConversationsLive.Chat
The chat view, its bubbles and the event→blocks seam move out of Show into a
shared function-component module so a second surface can render a
conversation the same way without copying 200 lines. No behaviour change.
